New _data submodule Also started with the specfile

+++ python_altinstall_packager/__init__.py
@@ -3,9 +3,10 @@
 
 """
 
-from json import loads as json_loads
 from logging import getLogger
 from pathlib import Path
+
+from . import _data as DATA
 
 try:
 	from jinja2 import Environment as Jinja2Environment
@@ -21,12 +22,72 @@ __version__ = '0.1.0.dev0'
 LOGGER = getLogger(__name__)
 
 
+class SpecfileTemplate(dict):
+	"""
+	
+	"""
+	
+	KNOWLEDGE_BASE_FILE = Path(__file__).parent / 'data' / 'distros.json'
+	TEMPLATE_NAME = 'specfile.jinja'
+	
+	def __getattr__(self, item):
+		"""
+
+		"""
+		
+		if item == 'client':
+			if docker_from_env is None:
+				raise ImportError('The "docker" package is required by {}'.format(type(self).__name__))
+			value = docker_from_env()
+		elif item == 'defaults':
+			if self.dist not in DATA.distributions:
+				raise ValueError('Missing "{}" distribution in knowledge base'.format(self.dist))
+			value = DATA.distributions[self.dist]
+			if 'dockerfile' in value:
+				self.update(value['dockerfile'])
+		elif item == 'packaging_tool':
+			if 'packaging_tool' not in self.defaults:
+				raise ValueError(
+					'Missing "packaging_tool" section for "{}" distribution in knowledge base'.format(self.dist))
+			value = self.defaults['packaging_tool']
+		elif item == 'template_file_path':
+			value = Path(__file__).parent / 'data' / self.TEMPLATE_NAME
+			if not value.is_file():
+				raise FileNotFoundError('Specfile template not found "{}"'.format(value))
+		else:
+			raise AttributeError(item)
+		
+		self.__setattr__(item, value)
+		return value
+	
+	def __init__(self, dist, python_version, /, **details):
+		"""
+		
+		"""
+		
+		super().__init__(details)
+		if isinstance(python_version, str):
+			python_version = python_version.split('.')
+		self['python_version'] = python_version
+	
+	def __str__(self):
+		"""
+
+		"""
+		
+		if Jinja2Environment is None:
+			raise ImportError('The "jinja2" package is required by {}'.format(type(self).__name__))
+		jinja_env = Jinja2Environment()
+		result = jinja_env.from_string(self.template_file_path.read_text())
+		return result.render(self)
+
+
 class DockerfileTemplate(dict):
 	"""
 
 	"""
 	
-	KNOWLEDGE_BASE_FILE = Path(__file__).parent / 'data' / 'knowledge_base.json'
+	KNOWLEDGE_BASE_FILE = Path(__file__).parent / 'data' / 'distros.json'
 	TAG_NAME = 'python_altinstall_packager'
 	TEMPLATE_NAMES = 'dockerfile_{}_template.jinja'
 	
@@ -45,7 +106,8 @@ class DockerfileTemplate(dict):
 		"""
 		
 		LOGGER.debug('Ignoring exception in context: %s(%s) | %s', exc_type, exc_val, exc_tb)
-		self.dockerfile.unlink(missing_ok=True)
+		if not self._keep_file:
+			self.dockerfile.unlink(missing_ok=True)
 	
 	def __getattr__(self, item):
 		"""
@@ -57,10 +119,9 @@ class DockerfileTemplate(dict):
 				raise ImportError('The "docker" package is required by {}'.format(type(self).__name__))
 			value = docker_from_env()
 		elif item == 'defaults':
-			value = json_loads(self.KNOWLEDGE_BASE_FILE.read_text())
-			if self.dist not in value:
+			if self.dist not in DATA.distributions:
 				raise ValueError('Missing "{}" distribution in knowledge base'.format(self.dist))
-			value = value[self.dist]
+			value = DATA.distributions[self.dist]
 			if 'dockerfile' in value:
 				self.update(value['dockerfile'])
 		elif item == 'packaging_tool':
@@ -77,7 +138,7 @@ class DockerfileTemplate(dict):
 		self.__setattr__(item, value)
 		return value
 	
-	def __init__(self, dist, python_version, root_dir=Path.cwd(), /, **details):
+	def __init__(self, dist, python_version, root_dir=Path.cwd(), keep_file=False, /, **details):
 		"""
 
 		"""
@@ -88,6 +149,7 @@ class DockerfileTemplate(dict):
 		if isinstance(python_version, str):
 			python_version = python_version.split('.')
 		self.python_version = python_version
+		self._keep_file = keep_file
 	
 	def __str__(self):
 		"""
@@ -134,4 +196,11 @@ class PythonAltinstallPackager:
 	
 	def test(self):
 		
-		return str(DockerfileTemplate('el9', '3.10.20'))
+		
+		# return str(SpecfileTemplate('el9', '3.10.20'))
+		
+		test_dir = Path.cwd() / 'testing'
+		test_dir.mkdir(exist_ok=True)
+		with DockerfileTemplate('el9', '3.10.20', test_dir, True) as dockerfile:
+			pass
+
